+++ CHANGELOG.md
@@ -7,6 +7,13 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [1.3.2] - 26.09.2021 
+
+## Fixed
+
+* Removed twig template cache and configured `twing-loader` as not cacheable.
+* Dropzone control attributes with correct values.
+
 ## [1.3.1] - 26.09.2021
 
 ## Fixed
@@ -230,7 +237,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 * Custom JS module support
 * Basic setup to publish project as NPM package
 
-[Unreleased]: https://github.com/bsi-software/bsi-cx-design-build/compare/1.3.1...HEAD
+[Unreleased]: https://github.com/bsi-software/bsi-cx-design-build/compare/1.3.2...HEAD
+
+[1.3.2]: https://github.com/bsi-software/bsi-cx-design-build/compare/1.3.1...1.3.2
 
 [1.3.1]: https://github.com/bsi-software/bsi-cx-design-build/compare/1.3.0...1.3.1